backend/pyproject.toml
[project]
name = "folio-enrich"
version = "0.1.0"
description = "Legal document annotation pipeline using the FOLIO ontology"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "folio-python>=0.1.5",
    "pyahocorasick>=2.0.0",
    "spacy>=3.7.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "markdown-it-py>=3.0.0",
    "sse-starlette>=2.0.0",
    "pyarrow>=15.0.0",
    "rdflib>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
